Fully functioning extraction script
Do all cities with population >1 millions & urban area greater than 100 km2 regardless of what SDEI has listed for the estimated warming (because they didn't do a whole lot of filtering it seems).

Also changed to iteratively filtering out 6-sigma outliers, but only on the mosaiced files and not applying the filter on the city scale; it will be easier to remove these on the post-hoc end rather than trying to re-run them here.  This also helps ensure that true outliers like the peak of mountain in ad-Dammam aren't erroniously removed