Base model selection based on residual distribution
Relying on traditional metrics of model fit like AIC or R2 caused residual outliers to have VERY large effects on our calculated value.  So rather than that, trying to use the model with the smaller range of residuals.  Lets see how that goes.

These outliers don't necessarily get caught by a simple 6-sigma filter, so if this doesn't work, it might be back to the drawing board.