Update threshold analysis
Including effects of differences in temperature on relative humidty.  Not sure this is the way to go, but it seemed safer than not.  I'm not really satisfied with this approach and think we'll probably resort to it being a theoretical discussion because the effects of trees and ET on humidity are big and we don't have any cities that have areas exposed to >35C at a seasonal scale.